AI_Analyst_Docker/requirements.txt
streamlit
pandas
plotly
python-dotenv
audio-recorder-streamlit
crewai
groq
requests